Fixed the 1-liner bug.
In generate_and_save(), np.genfromtxt reads files that have only 1 sample as a 1D array. All other files are read as 2D and this used to cause an iteration error because of the difference of dimensions. Now if we read a 1D array, we reshape it so it has 2D.